+++ projects/pip-webui2-documents/src/public_api.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of pip-webui2-documents
+ */
+
+export * from './lib/document/index';
+export * from './lib/document-edit/index';
+export * from './lib/document-list-edit/index';
+export * from './lib/document-list/index';