+++ src/app/app.module.ts
@@ -1,0 +1,67 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatToolbarModule, MatListModule, MatButtonModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
+import {
+  PipAppbarModule,
+  PipRightnavModule,
+  PipMediaModule,
+  PipSidenavModule,
+  PipMainLayoutModule,
+  PipRootLayoutModule,
+  PipSidenavExpanderModule
+} from 'pip-webui2-layouts-temp';
+import {
+  PipBreadcrumbModule,
+  PipNavIconModule,
+  PipPrimaryActionsModule,
+  PipSecondaryActionsModule,
+  PipNavModule,
+  PipNavHeaderModule,
+  PipNavMenuModule
+} from 'pip-webui2-nav-temp';
+
+import { PipThemesModule } from 'pip-webui2-themes-temp';
+
+import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { DocumentListExampleModule } from './document-list-example/document-list-example.module';
+import { DocumentListEditExampleModule } from './document-list-edit-example/document-list-edit-example.module';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    FlexLayoutModule,
+    MatToolbarModule, MatListModule, MatButtonModule,
+    TranslateModule.forRoot(),
+
+    PipThemesModule,
+    PipAppbarModule,
+    PipRightnavModule,
+    PipMediaModule,
+    PipSidenavModule,
+    PipMainLayoutModule,
+    PipRootLayoutModule,
+    PipSidenavExpanderModule,
+    PipBreadcrumbModule,
+    PipNavIconModule,
+    PipPrimaryActionsModule,
+    PipSecondaryActionsModule,
+    PipNavModule,
+    PipNavHeaderModule,
+    PipNavMenuModule,
+
+    AppRoutingModule,
+    DocumentListExampleModule,
+    DocumentListEditExampleModule
+  ],
+  bootstrap: [AppComponent],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
+})
+export class AppModule { }