+++ src/app/document-list-edit-example/document-list-edit-example.module.spec.ts
@@ -1,0 +1,13 @@
+import { DocumentListEditExampleModule } from './document-list-edit-example.module';
+
+describe('DocumentListEditExampleModule', () => {
+  let documentListEditExampleModule: DocumentListEditExampleModule;
+
+  beforeEach(() => {
+    documentListEditExampleModule = new DocumentListEditExampleModule();
+  });
+
+  it('should create an instance', () => {
+    expect(documentListEditExampleModule).toBeTruthy();
+  });
+});