+++ src/app/document-list-example/document-list-example.module.spec.ts
@@ -1,0 +1,13 @@
+import { DocumentListExampleModule } from './document-list-example.module';
+
+describe('DocumentListExampleModule', () => {
+  let documentListExampleModule: DocumentListExampleModule;
+
+  beforeEach(() => {
+    documentListExampleModule = new DocumentListExampleModule();
+  });
+
+  it('should create an instance', () => {
+    expect(documentListExampleModule).toBeTruthy();
+  });
+});