+++ src/document/index.ts
@@ -1,2 +1,0 @@
-export { PipDocumentComponent } from './document.component';
-export { PipDocumentModule } from './document.module';