+++ src/index.ts
@@ -1,4 +1,0 @@
-export * from './document/index';
-export * from './document-edit/index';
-export * from './document-list-edit/index';
-export * from './document-list/index';